requirements.txt
beautifulsoup4==4.12.3
docling==2.14.0
fastapi==0.115.6
fitz==0.0.1.dev2
gradio==5.9.1
json_repair==0.31.0
matplotlib==3.8.2
openai==1.58.1
pydantic==2.10.4
pymupdf4llm==0.0.17
python-dotenv==1.0.1
Requests==2.32.3
requests_toolbelt==1.0.0
spacy==3.7.5
tenacity==8.2.3
torch==2.5.1
uvicorn==0.34.0
wordcloud==1.9.4
